Changed simulation function

+++ DMpy/model.py
@@ -1241,7 +1241,7 @@ class DMModel():
             sim_static_theano.append(T.vector("sim_static_{0}".format(n)))
 
         for n, i in enumerate(sim_observation):
-            sim_observation_theano.append(T.vector("sim_observation_{0}".format(n)))
+            sim_observation_theano.append(T.vector("sim_observation_{0}".format(n), dtype='float64'))
 
         # sequences for scan should be in format (n_trials, n_subjects)
